module2/requirements.txt
google-api-python-client==2.149.0
google-auth==2.35.0
requests==2.32.3
boto3==1.35.74
